move DebugMenuView to binders package
